misc: chrome: rename files
